Refactored code with reusable buttons and Homepage div contents, Also added middleware and permission check for restricting user from the secure urls!

+++ src/books/views.py
@@ -1,12 +1,12 @@
+from django.core.exceptions import PermissionDenied
+from django.core.paginator import Paginator
+from django.db.models import Q
+from django.http import JsonResponse
 from django.shortcuts import render, redirect, get_object_or_404
 from django.views import View
+
 from src.books.forms import BookForm
 from src.books.models import Book
-
-from django.core.paginator import Paginator
-
-from django.http import JsonResponse
-from django.db.models import Q
 
 
 # Create your views here.
@@ -14,6 +14,7 @@ def hello(request):
     return render(request, '../../Project_B/templates/books/hello.html', {'name': 'Ayush'})
     # return render(request, 'hello.html')
     # return HttpResponse("Hello, %s!" % request.path)
+
 
 def search_books(request):
     query = request.GET.get('q', '').strip()
@@ -29,35 +30,39 @@ def search_books(request):
     ))
     return JsonResponse({'books': data})
 
+
 # List all books (Read)
 class BookListView(View):
-   def get(self, request):
-       limit = request.GET.get('limit',10)
-       try:
-           limit= int(limit)
-       except (ValueError, TypeError):
-           limit = 10 #fallback to default
+    # login_url = 'login_view'  # Optional: redirect URL for unauthenticated users
+
+    # redirect_field_name = 'next'  # Optional: default is 'next'
+
+    def get(self, request):
+        limit = request.GET.get('limit', 10)
+        try:
+            limit = int(limit)
+        except (ValueError, TypeError):
+            limit = 10  # fallback to default
+
+        # books = Book.objects.all()
+        books = Book.objects.all()
+
+        # Set Up Pagination
+        p = Paginator(Book.objects.all(), limit)
+        page = request.GET.get('page')
+        paginated_books = p.get_page(page)
+
+        return render(request, '../../Project_B/templates/books/book_list.html', {'books': books,
+                                                                                  'paginated_books': paginated_books,
+                                                                                  'limit': limit})
 
 
-
-
-
-       # books = Book.objects.all()
-       books = Book.objects.all()
-
-       # Set Up Pagination
-       p= Paginator(Book.objects.all(), limit)
-       page = request.GET.get('page')
-       paginated_books = p.get_page(page)
-
-       return render(request, '../../Project_B/templates/books/book_list.html', {'books': books,
-                                                                                 'paginated_books': paginated_books,'limit': limit})
-
-#View specific books(Read)
+# View specific books(Read)
 class BookDetailView(View):
     def get(self, request, uuid):
         book = get_object_or_404(Book, uuid=uuid)
-        return render(request,'../../Project_B/templates/books/book_detail_view.html',{'book': book} )
+        return render(request, '../../Project_B/templates/books/book_detail_view.html', {'book': book})
+
 
 # Create a book
 class BookCreateView(View):
@@ -65,38 +70,57 @@ class BookCreateView(View):
         form = BookForm()
         return render(request, '../../Project_B/templates/books/book_form.html', {'form': form})
 
-    def post(self,request):
+    def post(self, request):
         form = BookForm(request.POST)
         if form.is_valid():
             form.save()
             return redirect('book_list')
         return render(request, '../../Project_B/templates/books/book_form.html', {'form': form})
 
-#Update a book
+
+# Update a book
 class BookUpdateView(View):
-    def get(self,request,uuid):
+    def get(self, request, uuid):
+        print("before permission")
+        print(request.user.is_superuser)  # True means has all permissions
+        print(request.user.get_all_permissions())  # Shows all assigned permissions
+        if not request.user.has_perm('books.change_book'):
+            print("between permission")
+            raise PermissionDenied  # 403 Forbidden
+        print("After permission")
         book = get_object_or_404(Book, uuid=uuid)
         form = BookForm(instance=book)
         return render(request, '../../Project_B/templates/books/book_form.html', {'form': form})
 
-    def post(self,request,uuid):
+    def post(self, request, uuid):
+        print("before permission")
+        if not request.user.has_perm('books.change_book'):
+            print("between permission")
+            raise PermissionDenied  # 403 Forbidden
+        print("After permission")
         book = get_object_or_404(Book, uuid=uuid)
-        form = BookForm(request.POST,instance=book)
+        form = BookForm(request.POST, instance=book)
         if form.is_valid():
             form.save()
             return redirect('book_list')
-        return render(request, '../../Project_B/templates/books/book_form.html', {'form':form})
+        return render(request, '../../Project_B/templates/books/book_form.html', {'form': form})
 
-#Delete a book
+
+# Delete a book
 class BookDeleteView(View):
-    def get(self, request,uuid):
-        book = get_object_or_404(Book, uuid=uuid)
-        return render(request, '../../Project_B/templates/books/book_confirm_delete.html', {'book':book})
+    def get(self, request, uuid):
+        print("before permission")
+        if not request.user.has_perm('books.change_book'):
+            print("between permission")
+            raise PermissionDenied  # 403 Forbidden
+        print("After permission")
 
-    def post(self,request,uuid):
+        book = get_object_or_404(Book, uuid=uuid)
+        return render(request, '../../Project_B/templates/books/book_confirm_delete.html', {'book': book})
+
+    def post(self, request, uuid):
+        if not request.user.has_perm('books.change_book'):
+            raise PermissionDenied  # 403 Forbidden
         book = get_object_or_404(Book, uuid=uuid)
         book.delete()
         return redirect('book_list')
-
-
-
